# Conflicts:
#	cmake-build-debug/CMakeFiles/SBX_CON_LIB.dir/sbx_interface_def/Sbx_cpp_interface.cpp.o
#	cmake-build-debug/Testing/Temporary/LastTest.log
#	cmake-build-debug/libSBX_CON_LIB.a
#	sbx_interface_def/Sbx_cpp_interface.cpp

+++ sbx_interface_def/Sbx_c_connector.h
@@ -13,7 +13,7 @@ extern "C" {
 
 int c_isPointerToTaintedMem(void* pointer);
 
-void* c_fetch_pointer_from_offset(const unsigned long pointer_offset);
+void* c_fetch_pointer_from_offset(const unsigned int pointer_offset);
 
 int c_isTaintedPointerToTaintedMem(void* pointer);
